chore: 모듈 이름 변경 및 하위 패키지 구조로 리팩토링
- 모듈의 이름 변경
- 모듈에 맞도록 패키지 생성, 분리
