--- app/src/main/java/com/openstudy/carefull/chatbot/ChatMessage.kt
@@ -1,6 +1,0 @@
-package com.openstudy.carefull.chatbot
-
-data class ChatMessage(
-    val message: String,
-    val isUser: Boolean
-)
