--- app/src/main/java/com/openstudy/carefull/chatbot/ChatbotScreen.kt
@@ -1,94 +1,0 @@
-package com.openstudy.carefull.chatbot
-
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.material3.Button
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextField
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import com.openstudy.carefull.ui.theme.CarefullTheme
-
-@Composable
-fun ChatBotScreen(viewModel: ChatViewModel = ChatViewModel(), modifier: Modifier = Modifier) {
-    var userInput by remember { mutableStateOf("") }
-    
-    val messages = viewModel.chatMessages.reversed()
-
-    Column(modifier = modifier.fillMaxSize().padding(16.dp)) {
-        LazyColumn(
-            modifier = Modifier.weight(1f).fillMaxWidth(),
-            reverseLayout = true
-        ) {
-            items(messages) { message ->
-                ChatBubble(message)
-                Spacer(modifier = Modifier.height(8.dp))
-            }
-        }
-
-        Row(modifier = Modifier.fillMaxWidth()) {
-            TextField(
-                modifier = Modifier.weight(1f),
-                value = userInput,
-                onValueChange = { userInput = it },
-                placeholder = { Text("증상을 입력하세요") }
-            )
-            Spacer(modifier = Modifier.width(8.dp))
-            Button(onClick = {
-                viewModel.onUserMessage(userInput)
-                userInput = ""
-            }) {
-                Text("전송")
-            }
-        }
-    }
-}
-
-@Composable
-fun ChatBubble(message: ChatMessage) {
-    val alignment = if (message.isUser) Alignment.CenterEnd else Alignment.CenterStart
-    val bgColor = if (message.isUser) MaterialTheme.colorScheme.primary else MaterialTheme.colorScheme.surfaceVariant
-    val textColor = if (message.isUser) MaterialTheme.colorScheme.onPrimary else MaterialTheme.colorScheme.onSurface
-
-    Box(
-        modifier = Modifier.fillMaxWidth(),
-        contentAlignment = alignment
-    ) {
-        Surface(
-            color = bgColor,
-            shape = MaterialTheme.shapes.medium
-        ) {
-            Text(
-                text = message.message,
-                modifier = Modifier.padding(10.dp),
-                color = textColor
-            )
-        }
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun ChatBottScreenPreview() {
-    CarefullTheme {
-        ChatBotScreen()
-    }
-}
