--- app/src/main/java/com/openstudy/carefull/disease/DiseaseInfo.kt
@@ -1,6 +1,0 @@
-package com.openstudy.carefull.disease
-
-data class DiseaseInfo(
-    val name: String,
-    val description: String
-)
