--- app/src/main/java/com/openstudy/carefull/diseaseSearch/DiseaseInfo.kt
@@ -1,6 +1,0 @@
-package com.openstudy.carefull.diseaseSearch
-
-data class DiseaseInfo(
-    val name: String,
-    val description: String
-)
