--- app/src/main/java/com/openstudy/carefull/hospitalInfo/HospitalInfoScreen.kt
@@ -1,93 +1,0 @@
-package com.openstudy.carefull.hospitalInfo
-
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.viewinterop.AndroidView
-import com.naver.maps.geometry.LatLng
-import com.naver.maps.map.CameraPosition
-import com.naver.maps.map.CameraUpdate
-import com.naver.maps.map.MapView
-import com.openstudy.carefull.ui.theme.CarefullTheme
-
-@Composable
-fun HospitalInfoScreen() {
-
-    val context = LocalContext.current
-    val mapView = remember {
-        MapView(context).apply {
-            getMapAsync { naverMap ->
-                naverMap.cameraPosition = CameraPosition(
-                    LatLng(37.5665, 126.9780),
-                    14.0
-                )
-            }
-        }
-    }
-
-    var selectedHospitalType by remember { mutableStateOf("우수병원") }
-
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(16.dp)
-    ) {
-        Text(
-            text = "예상 병명",
-            style = MaterialTheme.typography.headlineSmall
-        )
-        Spacer(modifier = Modifier.height(8.dp))
-
-        Text(
-            text = "진료 과목",
-            style = MaterialTheme.typography.titleMedium
-        )
-        Spacer(modifier = Modifier.height(8.dp))
-
-        AndroidView(
-            factory = { mapView },
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(250.dp)
-                .clip(RoundedCornerShape(12.dp))
-        )
-
-        Spacer(modifier = Modifier.height(16.dp))
-
-        Text(
-            text = "병원 보기",
-            style = MaterialTheme.typography.titleMedium
-        )
-        Spacer(modifier = Modifier.height(8.dp))
-
-        Text(
-            text = "선택된 병원 종류: $selectedHospitalType",
-            style = MaterialTheme.typography.bodyLarge
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun HospitalInfoScreenPreview() {
-    CarefullTheme {
-        HospitalInfoScreen()
-    }
-}
