--- app/src/main/java/com/openstudy/carefull/medicineInfo/MedicineInfo.kt
@@ -1,7 +1,0 @@
-package com.openstudy.carefull.medicineInfo
-
-data class MedicineInfo(
-    val name: String,
-    val description: String,
-    val imageUrl: String
-)
