+++ data/src/main/java/com/openstudy/carefull/data/model/ChatBotMessageInfo.kt
@@ -1,0 +1,6 @@
+package com.openstudy.carefull.data.model
+
+data class ChatBotMessageInfo(
+    val message: String,
+    val isUser: Boolean
+)
