+++ data/src/main/java/com/openstudy/carefull/data/model/DiseaseInfo.kt
@@ -1,0 +1,6 @@
+package com.openstudy.carefull.data.model
+
+data class DiseaseInfo(
+    val name: String,
+    val description: String
+)
