+++ data/src/main/java/com/openstudy/carefull/data/model/MedicineInfo.kt
@@ -1,0 +1,7 @@
+package com.openstudy.carefull.data.model
+
+data class MedicineInfo(
+    val name: String,
+    val description: String,
+    val imageUrl: String
+)
