+++ shared/src/main/java/com/openstudy/carefull/shared/constants/example.kt
@@ -1,0 +1,4 @@
+package com.openstudy.carefull.shared.constants
+
+class example {
+}
